baremetal: Use machine-os-images to get CoreOS
The machine-os-images container supplies the latest ISO and PXE images
for pre-provisioning from within the release payload. Use this instead
of downloading CoreOS images. Extract the images to the same location
(and with the same names) in the image cache volume that we were using
previously for IPA images. This change is transparent to the tasks using
the images.