OCPBUGS-1824: Bind agent config scripts to assisted-service-pod
If the assisted-service pod restarts (e.g. because the startup times
out), we want everything that depends on it to try again. The PartOf
directive evidently isn't sufficient to do that, since after the pod
service restarts the subsequent services (such as
create-cluster-and-infraenv) report that they won't start due to a
failed dependency.

Change from PartOf to BindsTo to ensure that once the pod succeeds in
starting up, its dependencies will also start.