Temporarily use identical tokens for all roles
Until assisted-service supports different tokens for the agent, user,
and watcher roles, use the same token for all three. Pass the necessary
environment variables and HTTP Headers for both auth methods. This
should pass CI with the existing assisted-service, and allow the support
for the new authorization scheme to land in assisted-service. After
that, this patch should be reverted.